processing/requirements.txt
pyspark==3.5.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1